Move TriviaSession to trivia.ts, exclude triviaLists, and document /trivia following code review
