+++ .gitignore
@@ -66,3 +66,6 @@ typings/
 
 # dotenv environment variables file
 .env
+
+# trivia lists
+triviaLists/
