+++ README.md
@@ -1,6 +1,6 @@
 # IveBot
 
-The bot that created the iPhone X. It's strictly private. You may run it locally,  but it's not a useful bot to others and tailored for a specific need.
+The bot that created the iPhone X. It's strictly private. You may run it locally, but it's not a useful bot to others and tailored for a specific need.
 
 **Requires Node.js 8.5.0 or higher.**
 
@@ -24,6 +24,7 @@ It's planned to have some nifty commands like /assistant which directly communic
 - `/randomword`
 - `/choose`
 - `/reverse`
+- `/trivia`
 - `/8ball`
 - `/repeat`
 - `/calculate`
@@ -78,6 +79,8 @@ It's planned to have some nifty commands like /assistant which directly communic
 ## Configuration
 
 Set up a MongoDB instance and note its URL. You can set it to store its data in `database` within this folder (you must first make the folder before starting MongoDB)
+
+Get the trivia lists zip from [here](https://siasky.net/nAHYx0Qe7NFag-RuMZTSGizq5ral6Q6m6BZrSHKzzx7r_g) and extract it so that the triviaLists folder is in the top-level directory.
 
 Make a file named `config.json5` in the top-level directory. It should be something like this:
