+++ server/index.ts
@@ -74,6 +74,8 @@ MongoClient.connect(mongoURL === 'dotenv' ? process.env.MONGO_URL : mongoURL, {
         if (!Object.keys(commands).length) return
         // ..register the commands.
         Object.keys(commands).forEach((commandName: string) => {
+          // exclude TriviaSession from commands
+          if (commandName === 'TriviaSession') { return }
           const command = commands[commandName]
           commandParser.registerCommand(command)
         })
